patch formatting improvements
